Split `PurchasesTests` step 1: extracted `BasePurchasesTests`
